move to correct location
