+++ app/diagnostic/webhook/check.go
@@ -25,7 +25,7 @@ import (
 	config "github.com/cloudzero/cloudzero-agent/app/config/validator"
 	"github.com/cloudzero/cloudzero-agent/app/diagnostic"
 	logging "github.com/cloudzero/cloudzero-agent/app/logging/validator"
-	"github.com/cloudzero/cloudzero-agent/app/status"
+	"github.com/cloudzero/cloudzero-agent/app/types/status"
 )
 
 const DiagnosticInsightsIngress = config.DiagnosticInsightsIngress
