+++ app/telemetry/post.go
@@ -14,9 +14,9 @@ import (
 
 	"google.golang.org/protobuf/proto"
 
-	"github.com/cloudzero/cloudzero-agent/app/config/validator"
-	"github.com/cloudzero/cloudzero-agent/app/http/client"
-	pb "github.com/cloudzero/cloudzero-agent/app/status"
+	config "github.com/cloudzero/cloudzero-agent/app/config/validator"
+	http "github.com/cloudzero/cloudzero-agent/app/http/client"
+	pb "github.com/cloudzero/cloudzero-agent/app/types/status"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 )
